Fixed all domain entity data mismatch types and relationships

+++ services/src/main/java/za/co/entelect/bootcamp/flash/services/SupplierPaymentDTOAdapter.java
@@ -3,6 +3,8 @@ package za.co.entelect.bootcamp.flash.services;
 import za.co.entelect.bootcamp.flash.domain.Supplier;
 import za.co.entelect.bootcamp.flash.domain.SupplierPayment;
 
+import java.math.BigDecimal;
+
 /**
  * Created by byron.dinkelmann on 2017/01/16.
  */
@@ -38,12 +40,12 @@ public class SupplierPaymentDTOAdapter extends SupplierPaymentDTO {
     }
 
     @Override
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return this.supplierPayment.getTotal();
     }
 
     @Override
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         super.setAmount(amount);
     }
